move AnimatedDots be reusable
